部署准备：修复考勤查看、自动创建学生、QR码、服务器配置

+++ src/main/java/com/example/attendance/dto/CourseOption.java
@@ -6,9 +6,8 @@ import java.time.LocalTime;
  * 课程下拉选项：用于打卡与筛选。
  */
 public record CourseOption(
-        Integer id,
+        Long id,
         String name,
         LocalTime startTime
 ) {
 }
-
